requirements.txt
Django==3.2.15
django-environ==0.9.0

